+++ src/pos_controller.hpp
@@ -12,13 +12,13 @@ public:
   /// \param Ki - The integral gain for a PID controller
   /// \param Kd - The derivative gain for a PID controller
   /// \param Kff - The feed forward gain
-  PositionController(double Kp, double Ki, double Kd, double Kff);
+  PositionController(float Kp, float Ki, float Kd, float Kff);
 
   /// \brief Pump controller
   /// \param setpoint - The set point position in radians
   /// \param actual - The actual motor shaft position in radians
   /// \param shaft_vel - The motor shaft velocity used in derivative control
-  double pump_controller(double setpoint, double actual, float shaft_vel);
+  float pump_controller(float setpoint, float actual, float shaft_vel);
 
   /// \brief Set the status of feed forward control in the class
   /// \param enable - Sets if feed forward term is used in control
@@ -26,41 +26,41 @@ public:
 
   /// \brief Set the value of the integral control error clamp
   /// \param clamp_val - The value that the integral error is clamped at
-  void set_i_clamp_val(double clamp_val);
+  void set_i_clamp_val(float clamp_val);
 
   /// \brief Set the value of the command clamp
   /// \param clamp_val - The value that the command output is clamped at
-  void set_u_clamp_val(double clamp_val);
+  void set_u_clamp_val(float clamp_val);
 
 private:
   /// \brief Proportional gain
-  double Kp_;
+  float Kp_;
   /// \brief Integral gain
-  double Ki_;
+  float Ki_;
   /// \brief Derivative gain
-  double Kd_;
+  float Kd_;
   /// \brief Feed forward gain
-  double Kff_;
+  float Kff_;
   /// \brief Boolean enabling integral control
   bool Icntrl_;
   /// \brief Boolean enabling derivative control
   bool Dcntrl_;
   /// \brief Error
-  double err_;
+  float err_;
   /// \brief Integrated error
-  double err_int_;
+  float err_int_;
   /// \brief Derivative error
-  double err_der_;
+  float err_der_;
   /// \brief Previous error
-  double err_prev_;
+  float err_prev_;
   /// \brief Integral error clamping value
-  double i_clamp_val_;
+  float i_clamp_val_;
   /// \brief Command clamping value
-  double u_clamp_val_;
+  float u_clamp_val_;
   /// \brief Boolean enabling gravity compensation control
   bool gvty_fwd_enable_;
   /// \brief Gravity compensation term
-  double gvty_term_;
+  float gvty_term_;
 };
 
 #endif
